bardock/bardockpro: Fix calling CheckDigPinConnectivity.
We need to update this library to match the requirements
(library exports) when switching to BluetoothQti.

So the newer lib is taken from the casuara firmware, as the
remaining blobs taken from. But some special versions of our
beloved phone doesn't handle the call to
CheckDigPinConnectivity correctly, but for most of the devices
around this call succeeds. In stock version
this call doesn't exists so I assume, this isn't really needed.

Let's manual patch the file
android.hardware.bluetooth@1.0-impl-qti.so.
to not check the result of the method call by replacing the
instruction with a nop command.

Change-Id: I2a95c50cce5d106244b01f59fd08f43513c04077