bardock/bardockpro: Patch qcril to support only RadioConfig::1.0 interface.
For experimental firmware. Currently doesn't affect normal
use case which is restricted to rild and not qcrild.
Only one drawback was encountered:
Modem doesn't support HAL_COMMAND_PREFERRED_DATA which is used
on android.hardware.radio.config@1.1::IRadioConfig interfaces.
On 1.0 this is solved in an different manner.
This is the reason why the libril-qc-hal-qmi.so is linked
to a modified hal stub.

Change-Id: I0fc92caa511d87fa8b431146d31567cc6ee2fa8c